Split cooler attribute tests
